Fixing v1 release around referencing validate.js
In moving this action publicly, the approach of sourcing the JavaScript
module used for validating the issue form broke, which I suspected it
would.

Instead of hardcoding the path, I need to use the environment
variable referencing the location where the action was cloned.  There is
a problem as I cannot use expressions in the default value of action
variables and unsure if we should evaluate an expression passed in from
input.